+++ react/src/components/Pagination/index.js
@@ -1,0 +1,1 @@
+export { Pagination, PaginationPrev, PaginationNext, PaginationItem, PaginationEllipsis, default } from './Pagination';