+++ react/src/index.js
@@ -10,4 +10,5 @@ export { Avatar, AvatarStack, AvatarStatus } from './components/Avatar';
 export { Alert, AlertBtnPrimary, AlertBtnSecondary } from './components/Alert';
 export { Tooltip, TooltipHost } from './components/Tooltip';
 export { Tabs, Tab, TabPanel } from './components/Tabs';
+export { Pagination, PaginationPrev, PaginationNext, PaginationItem, PaginationEllipsis } from './components/Pagination';
 export { default } from './components/Button';